requirements.txt
numpy
pandas
flask
flask_cors
openai-whisper
deepmultilingualpunctuation
pyaudio